requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
PYJWT
PYMYSQL
Python-Dotenv
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cv==1.0.0
cycler==0.10.0
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fer==20.1.3
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
mtcnn==0.1.0
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.3
Pillow==8.2.0
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.2
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1